Add TemperatureConverter to lab3 and update Program.cs
In this commit, I've merged the TemperatureConverter feature, previously in a separate project, into the lab3 project. This change simplifies demonstration of the code producing the TemperatureConverter.dll. I also revised Program.cs to reflect a new implementation, using reflection to invoke the ConvertCelsiusToFahrenheit method from the TemperatureConverter class. Now it takes temperature in Celsius as input and prints out the corresponding temperature in Fahrenheit.